+++ packages/batch/src/index.ts
@@ -6,6 +6,6 @@
  * @module
  */
 
-export { batchConcurrent } from './batch/index.js';
+export { Batch } from './batch/index.js';
 export { BatchError } from './errors/index.js';
-export type { BatchHooksInterface, BatchOptionsInterface, BatchStatsType } from './interfaces/index.js';
+export type { BatchStatsType } from './interfaces/index.js';